Moved data point struct and implementation into separate file.
